apply copilot's comments
