+++ sdk/provisioning/Azure.Provisioning/src/Expressions/AST/Expressions/BicepExpression.cs
