+++ sdk/provisioning/Azure.Provisioning/src/Expressions/AST/Expressions/InterpolatedStringExpression.cs
