+++ sdk/provisioning/Azure.Provisioning/src/Expressions/AST/Expressions/SafeMemberExpression.cs
@@ -3,14 +3,6 @@
 
 namespace Azure.Provisioning.Expressions;
 
-public class MemberExpression(BicepExpression value, string member) : BicepExpression
-{
-    public BicepExpression Value { get; } = value;
-    public string Member { get; } = member;
-    internal override BicepWriter Write(BicepWriter writer) =>
- writer.Append(Value).Append('.').Append(Member);
-}
-
 public class SafeMemberExpression(BicepExpression value, string member) : BicepExpression
 {
     public BicepExpression Value { get; } = value;
