+++ sdk/provisioning/Azure.Provisioning/src/Expressions/AST/Expressions/StringLiteralExpression.cs
@@ -6,8 +6,7 @@ namespace Azure.Provisioning.Expressions;
 public class StringLiteralExpression(string value) : LiteralExpression(value)
 {
     public new string Value { get => (string)base.Value!; }
-    internal override BicepWriter Write(BicepWriter writer) =>
- (Value == null) ?
- writer.Append("null") :
-   writer.Append('\'').AppendEscaped(Value).Append('\'');
+    internal override BicepWriter Write(BicepWriter writer) => Value == null ?
+        writer.Append("null") :
+        writer.Append('\'').AppendEscaped(Value).Append('\'');
 }
