+++ sdk/provisioning/Azure.Provisioning/src/Expressions/AST/Expressions/UnaryBicepOperator.cs
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Azure.Provisioning.Expressions;
+
+public enum UnaryBicepOperator { Not, Negate, SuppressNull }
